+++ src/components/nav.html
@@ -8,7 +8,7 @@
     <div class="collapse navbar-collapse" id="navbarNav">
         <ul class="navbar-nav ml-auto">
             <li class="nav-item dashboard-link">
-                <a  class="nav-link" href="/dashboard">Dashboard</a>
+                <a class="nav-link" href="/dashboard">Dashboard</a>
             </li>
             <li data-netlify-identity-menu></li>
         </ul>
