+++ src/pages/professor/dashboard.html
@@ -1,0 +1,126 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <!-- Global site tag (gtag.js) - Google Analytics -->
+    <script async src="https://www.googletagmanager.com/gtag/js?id=UA-128850591-1"></script>
+    <script>
+        if (document.location.hostname === 'www.insiight.ca') {
+            window.dataLayer = window.dataLayer || [];
+            function gtag(){dataLayer.push(arguments);}
+            gtag('js', new Date());
+
+            gtag('config', 'UA-128850591-1');
+        }
+    </script>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <title>InSiight | Dashboard</title>
+    <link rel="shortcut icon" type="image/png" href="${require('../../assets/logo_transparent.png')}" />
+    <!-- Twitter Card data -->
+    <meta name="twitter:card" value="InSiight">
+    <meta name="Description" content="Transforming professor-student feedback interactions for a better educational experience."/>
+    <!-- Open Graph data -->
+    <meta property="og:title" content="InSiight" />
+    <meta property="og:type" content="article" />
+    <meta property="og:url" content="https://www.InSiight.ca" />
+    <meta property="og:image" content="${require('../../assets/logo_transparent.png')}" />
+    <meta property="og:description" content="Transforming professor-student feedback interactions for a better educational experience." />
+    <!-- Mobile header color for Chrome, Firefox OS and Opera -->
+    <meta name="theme-color" content="#000000">
+    <!-- Mobile header color Windows Phone -->
+    <meta name="msapplication-navbutton-color" content="#000000">
+    <!-- Mobile header color for iOS Safari (supports black, black-translucent or default) -->
+    <meta name="apple-mobile-web-app-status-bar-style" content="default">
+    <link href="https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700" rel="stylesheet">
+    <!-- Bootstrap 4 -->
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
+    <!-- Netlify Identity Widget -->
+    <script type="text/javascript" src="https://identity.netlify.com/v1/netlify-identity-widget.js"></script>
+    <!-- Survey.js CSS -->
+    <link href="https://surveyjs.azureedge.net/1.0.56/survey.css" type="text/css" rel="stylesheet" />
+</head>
+<body>
+    <div class="container">
+        ${require('../../components/nav.html')}
+    </div>
+    <main class="container-fluid dashboard-page p-0">
+        <!-- Sidebar -->
+        <div class="sidebar">
+            <div class="courses">
+            </div>
+            <div class="bottom-section">
+                <button data-toggle="modal" data-target="#add-course-modal">
+                    <img src="${require('../../assets/plus.svg')}" alt="Add course" />
+                </button>
+            </div>
+        </div>
+
+        <!-- Survey -->
+        <div class="survey">
+            <div class="professor-survey-options">
+                <div class="row">
+                    <div class="col-sm-4">
+                        <div class="card">
+                            <div class="card-body">
+                                <h5 class="card-title">Send a survey</h5>
+                                <p class="card-text">Start receiving feedback from your students in your course by sending them a survey!</p>
+                                <a href="#" class="btn btn-primary">Send survey</a>
+                            </div>
+                        </div>
+                    </div>
+                    <div class="col-sm-4">
+                        <div class="card">
+                            <div class="card-body">
+                                <h5 class="card-title">View title treatment</h5>
+                                <p class="card-text">With supporting text below as a natural lead-in to additional content.</p>
+                                <a href="#" class="btn btn-primary">Go somewhere</a>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+        </div>
+            <div class="survey-container"></div>
+        </div>
+    </main>
+
+    <!-- Add course modal -->
+    <div class="modal fade" id="add-course-modal" tabindex="-1" role="dialog" aria-labelledby="add-course-modal-title" aria-hidden="true">
+        <div class="modal-dialog modal-dialog-centered" role="document">
+          <div class="modal-content">
+            <div class="modal-header">
+              <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                <span aria-hidden="true">&times;</span>
+              </button>
+            </div>
+            <div class="modal-body">
+                <form id="add-course-form">
+                    <div class="professor-form input-block">
+                        <h5>Join your school and add your course</h5>
+                        <div>
+                            <label for="school-name-input">School name</label>
+                            <input id="school-name-input" name="schoolName" type="text" />
+                        </div>
+                        <div>
+                            <label for="course-name-input">Course name</label>
+                            <input id="course-name-input" name="courseName" type="text" />
+                        </div>
+                    </div>
+                </form>
+            </div>
+            <div class="modal-footer">
+                <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+                <button type="button" class="btn primary-color add-course-button" form="add-course-form">Add course</button>
+            </div>
+          </div>
+        </div>
+    </div>
+
+    <!-- Pick role modal -->
+    ${require('../../components/role-modal.html')}
+
+    <!-- Bootstrap 4 & jQuery scripts -->
+    ${require('../../components/bootstrap-scripts.html')}
+    <!-- Survey.js with jQuery -->
+    <script src="https://surveyjs.azureedge.net/1.0.56/survey.jquery.js"></script>
+</body>
+</html>
