+++ src/pages/student/dashboard.html
@@ -55,9 +55,36 @@
             </div>
         </div>
 
-        <!-- Main content -->
-        <div class="main">
-            <div id="surveyContainer"></div>
+        <!-- Survey -->
+        <div class="survey">
+            <div class="professor-survey-options">
+                <div class="row">
+                    <div class="col-sm-4">
+                        <div class="card">
+                            <div class="card-body">
+                                <h5 class="card-title">Send a survey</h5>
+                                <p class="card-text">Start receiving feedback from your students in your course by sending them a survey!</p>
+                                <a href="#" class="btn btn-primary">Send survey</a>
+                            </div>
+                        </div>
+                    </div>
+                    <div class="col-sm-4">
+                        <div class="card">
+                            <div class="card-body">
+                                <h5 class="card-title">View title treatment</h5>
+                                <p class="card-text">With supporting text below as a natural lead-in to additional content.</p>
+                                <a href="#" class="btn btn-primary">Go somewhere</a>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+        </div>
+            <div class="survey-container"></div>
+        </div>
+
+        <!-- Discussion thread -->
+        <div class="discussion-thread">
+            <h3>Discussion Thread</h3>
         </div>
     </main>
 
@@ -76,17 +103,6 @@
                         <h5>Find your course with a JOIN CODE</h5>
                         <label for="course-code-input">Every course has a 4-digit join code that your professor gave you.</label>
                         <input id="course-code-input" name="courseCode" type="text" maxlength="4" />
-                    </div>
-                    <div class="professor-form input-block">
-                        <h5>Join your school and add your course</h5>
-                        <div>
-                            <label for="school-name-input">School name</label>
-                            <input id="school-name-input" name="schoolName" type="text" />
-                        </div>
-                        <div>
-                            <label for="course-name-input">Course name</label>
-                            <input id="course-name-input" name="courseName" type="text" />
-                        </div>
                     </div>
                 </form>
             </div>
