+++ src/styles/dashboard.scss
@@ -1,7 +1,7 @@
 .dashboard-page {
 
     height: calc(100vh - 58px); // remove height from navbar
-    background-color: #CCC;
+    background-color: white;
     display: flex;
 
     .sidebar {
@@ -10,7 +10,6 @@
         display: flex;
         flex-direction: column;
         background-color: #AAA;
-        border-top: 2px solid #83DFDB;
 
         .courses {
             height: 100%;
@@ -18,7 +17,21 @@
         }
 
         .course-info {
+            width: 100%;
             background: paleturquoise;
+            margin: 0 0 10px 0;
+
+            &:hover {
+                cursor: pointer;
+            }
+            &:focus,
+            &:checked {
+                background-color: #83DFDB;
+            }
+            
+            h5 {
+                margin-top: 0.5rem;
+            }
         }
 
         .bottom-section {
@@ -46,11 +59,13 @@
         }
     }
 
-    .main {
-        width: 100%;
-        height: 100%;
-        display: flex;
-        border-top: 2px solid #83DFDB;
+    .discussion-thread {
+        width: 30%;
+        border: 1px solid black;
+
+        h3 {
+            margin: 1rem 2rem;
+        }
     }
 }
