+++ src/styles/survey.scss
@@ -1,24 +1,44 @@
-#surveyContainer{
-    width: 100%;
-    max-height: 100%;
-    overflow-y: auto;
+.survey {
+    // display: flex;
+    width: 60%;
+    height: 100%;
     background-color: white;
 
-    .sv_main {
-        font-size: 1rem;
+    .professor-survey-options {
+        height: 100%;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+    
+        .row {
+            justify-content: center;
+        }
+    }
 
-        .sv_body {
-            border-top: 0;
-            padding: 0;
+    .survey-container {
+        width: 100%;
+        max-height: 100%;
+        overflow-y: auto;
 
-            .sv_q {
-                padding: 0.5rem 3rem;
-            }
+        .sv_main {
+            font-size: 1rem;
 
-            .sv_complete_btn {
-                margin-right: 15px;
+            .sv_body {
+                border-top: 0;
+                padding: 0;
+
+                .sv_page_title {
+                    padding: 1rem 3rem 0 3rem;
+                }
+
+                .sv_q {
+                    padding: 0.5rem 3rem;
+                }
+
+                .sv_complete_btn {
+                    margin-right: 15px;
+                }
             }
         }
     }
 }
-
